variable height jump/enemy patrol/ healthbar/knockback when hit
added a health bar that changes when hit by the enemy. enemy will now move back and forth. when player collides with enemy he will be knocked back (not fully working).  moved movement into fixed time and made small change so the knockback works( code is still the same its just in a if so it will disable when being knocked back). added vaiable height jump and removed the jump method(looked useless lmk if it wasnt and i will add it back.